programs/server/lib/pyproject.toml
[project]
name = "server-lib"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "shared-lib",
    "torchaudio>=2.6.0",
    "numpy>=2.4.2",
    "chatterbox-tts>=0.1.3",
]

[tool.uv.sources]
shared-lib = { path = "../../../lib", editable = true }

[tool.uv.extra-build-dependencies]
pkuseg = ["numpy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

